🎨 (#8) 전체 구성을 TableView에서 CollectionView로 전환
